Created reusable gallery component
