+++ src/Components/styles/GalleryStyles.css
@@ -1,6 +1,3 @@
-.galleryPage {
-}
-
 img {
   -moz-user-select: none;
   -webkit-user-select: none;
